Fix Chaquopy 17 API (android.AndroidPlatform) and Long/Int mismatch

+++ app/src/main/java/com/toolkit/app/PythonScreen.kt
@@ -119,7 +119,7 @@ fun PythonScreen(onBack: () -> Unit) {
             val dest = File(dir, safeName)
             dest.writeBytes(bytes)
             fileName = safeName
-            fileSize = bytes.size
+            fileSize = bytes.size.toLong()
             filePath = dest.absolutePath
             preview = content.lines().take(8).joinToString("\n").ifBlank { "(пустой файл)" }
             deps = scanDeps(content)
